+++ modules/invenio-indexer/invenio_indexer/ext.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 #
 # This file is part of Invenio.
-# Copyright (C) 2016-2018 CERN.
+# Copyright (C) 2016-2022 CERN.
 # Copyright (C) 2016 TIND.
 #
 # Invenio is free software; you can redistribute it and/or modify it
@@ -9,14 +9,12 @@
 
 """Flask exension for Invenio-Indexer."""
 
-from __future__ import absolute_import, print_function
-
-import six
 from flask import current_app
 from werkzeug.utils import cached_property, import_string
 
 from . import config
 from .cli import run  # noqa
+from .registry import IndexerRegistry
 from .signals import before_record_index
 
 
@@ -37,11 +35,12 @@ class InvenioIndexer(object):
         :param app: The Flask application.
         """
         self.init_config(app)
-        app.extensions['invenio-indexer'] = self
+        self.registry = IndexerRegistry()
+        app.extensions["invenio-indexer"] = self
 
-        hooks = app.config.get('INDEXER_BEFORE_INDEX_HOOKS', [])
+        hooks = app.config.get("INDEXER_BEFORE_INDEX_HOOKS", [])
         for hook in hooks:
-            if isinstance(hook, six.string_types):
+            if isinstance(hook, str):
                 hook = import_string(hook)
             before_record_index.connect_via(app)(hook)
 
@@ -51,10 +50,10 @@ class InvenioIndexer(object):
         :param app: The Flask application.
         """
         for k in dir(config):
-            if k.startswith('INDEXER_'):
+            if k.startswith("INDEXER_"):
                 app.config.setdefault(k, getattr(config, k))
 
     @cached_property
     def record_to_index(self):
         """Import the configurable 'record_to_index' function."""
-        return import_string(current_app.config.get('INDEXER_RECORD_TO_INDEX'))
+        return import_string(current_app.config.get("INDEXER_RECORD_TO_INDEX"))